Move kafka-mapper commands to 'experimental'
