+++ src/cmd/exp/kafkamapper/kafkamapper.go
@@ -1,19 +1,17 @@
 package kafkamapper
 
 import (
-	"github.com/otterize/otterize-cli/src/cmd/groups"
-	"github.com/otterize/otterize-cli/src/cmd/kafkamapper/export"
-	"github.com/otterize/otterize-cli/src/cmd/kafkamapper/list"
-	"github.com/otterize/otterize-cli/src/cmd/kafkamapper/visualize"
+	"github.com/otterize/otterize-cli/src/cmd/exp/kafkamapper/export"
+	"github.com/otterize/otterize-cli/src/cmd/exp/kafkamapper/list"
+	"github.com/otterize/otterize-cli/src/cmd/exp/kafkamapper/visualize"
 	"github.com/spf13/cobra"
 )
 
 var KafkaMapperCmd = &cobra.Command{
 	Use:     "kafka-mapper",
-	GroupID: groups.OSSGroup.ID,
 	Aliases: []string{"kafka"},
 	Short:   "Map kafka topic-level access",
-	Hidden:  true,
+	//Hidden:  true,
 }
 
 func init() {
